+++ src/pages/mat-mangmnt/menu/Receipt/custjobwork/sheets and others/SheetsClosedRVList.js
@@ -3,6 +3,7 @@ import { dateToShort } from "../../../../../../utils";
 import BootstrapTable from "react-bootstrap-table-next";
 import paginationFactory from "react-bootstrap-table2-paginator";
 import { useNavigate } from "react-router-dom";
+import { toast } from "react-toastify";
 
 const { getRequest, postRequest } = require("../../../../../api/apiinstance");
 const { endpoints } = require("../../../../../api/constants");
@@ -42,9 +43,13 @@ function SheetsClosedRVList() {
   const openButtonClick = () => {
     //console.log("data = ", data);
     //console.log("button click : ");
-    nav("/materialmanagement/receipt/openbuttonclosedsheetunit", {
-      state: { id: data.RvID },
-    });
+    if (data && data.RvID !== "") {
+      nav("/materialmanagement/receipt/openbuttonclosedsheetunit", {
+        state: { id: data.RvID },
+      });
+    } else {
+      toast.error("Select Customer");
+    }
   };
 
   const selectRow = {
